refactor: create Schedule module and clean up
