+++ packages/aztecs-sdl-image/examples/Sprite.hs
@@ -46,7 +46,7 @@ app =
   schedule SDL.setup
     >>> schedule IMG.setup
     >>> schedule setup
-    >>> S.forever
+    >>> forever
       ( schedule IMG.load
           >>> schedule SDL.update
           >>> schedule IMG.draw
