+++ packages/aztecs/aztecs.cabal
@@ -26,6 +26,7 @@ library
         Data.Aztecs.Component
         Data.Aztecs.Entity
         Data.Aztecs.Query
+        Data.Aztecs.Schedule
         Data.Aztecs.Storage
         Data.Aztecs.System
         Data.Aztecs.View
