+++ packages/aztecs/src/Data/Aztecs.hs
@@ -75,6 +75,7 @@ import Data.Aztecs.Access (Access, runAccess)
 import Data.Aztecs.Component (Component (..))
 import Data.Aztecs.Entity (EntityID)
 import Data.Aztecs.Query (Query, QueryFilter, with, without)
-import Data.Aztecs.System (Schedule, System,SystemT,  forever, runSchedule, runSchedule_, schedule)
+import Data.Aztecs.Schedule (Schedule, forever, runSchedule, runSchedule_, schedule)
+import Data.Aztecs.System (System, SystemT)
 import Data.Aztecs.World (World)
 import Data.Aztecs.World.Archetype (Bundle, bundle)
